requirements.txt
fastapi
uvicorn
celery
redis
python-dotenv
requests
gitpython          
langchain          
langchain-openai   
shutil            